Merging Release/8.0 into develop after clang reformatting

+++ src/solver/variable/storage/averagedata.h
@@ -77,8 +77,8 @@ public:
     mutable double allYears; // FIX MEEE - Remove the mutable as soon as possible
     unsigned int nbYearsCapacity;
 
-    std::vector<int> yearsWeight;
-    int yearsWeightSum;
+    std::vector<float> yearsWeight;
+    float yearsWeightSum;
 
 }; // class AverageData
