+++ src/solver/variable/storage/stdDeviation.h
@@ -306,8 +306,8 @@ private:
     }
 
 private:
-    std::vector<int> yearsWeight;
-    int yearsWeightSum;
+    std::vector<float> yearsWeight;
+    float yearsWeightSum;
 
 }; // class StdDeviation
